+++ config/routes.rb
@@ -4,6 +4,7 @@ Rails.application.routes.draw do
   get "health_check" => "application#health_check"
   root to: "infrastructures#index"
 
+  resources :dalmatian_builds, only: [:index]
   resources :infrastructures, only: [:show, :index] do
     resources :environment_variables, only: [:new, :create, :destroy, :index] do
       collection do